Convert Dashboard Controller into Form.
