+++ modules/datastore/datastore.routing.yml
@@ -156,12 +156,11 @@ datastore.dkan_resource_settings:
     _permission: 'administer site configuration'
 
 datastore.datasets_import_status_dashboard:
-  path: '/admin/dkan/dataset-status/{harvestId}'
+  path: '/admin/dkan/dataset-status'
   methods: [ GET ]
   defaults:
-    _controller: \Drupal\datastore\Controller\DashboardController::buildDatasetsImportStatusTable
-    _title_callback: \Drupal\datastore\Controller\DashboardController::buildDatasetsImportStatusTitle
-    harvestId: NULL
+    _title: 'Datastore Import Status'
+    _form: 'Drupal\datastore\Form\DashboardForm'
   requirements:
     _permission: 'dkan.harvest.dashboard'
   options:
