+++ modules/datastore/src/Service/ResourceLocalizer.php
@@ -48,7 +48,7 @@ class ResourceLocalizer {
    */
   public function localize($identifier, $version = NULL) {
     $resource = $this->getResourceSource($identifier, $version);
-    if ($resource) {
+    if (isset($resource) && !file_exists(UrlHostTokenResolver::resolveFilePath($resource->getFilePath()))) {
       $ff = $this->getFileFetcher($resource);
       return $ff->run();
     }
