+++ modules/harvest/src/DashboardController.php
@@ -68,7 +68,7 @@ class DashboardController {
    * Private.
    */
   private function buildHarvestRow(string $harvestId, string $runId, $info) {
-    $url = Url::fromRoute("datastore.datasets_import_status_dashboard", ['harvestId' => $harvestId]);
+    $url = Url::fromRoute('datastore.datasets_import_status_dashboard', ['harvest_id' => $harvestId]);
 
     return [
       'harvest_link' => Link::fromTextAndUrl($harvestId, $url),
